New version of the AlCa matrix.
The Alca matrix has a new structure in order to take into account the data processing scenarios, which differ between Run1 and Run2, as well as the PD definition, which is bound to HLT menu changing from cosmics to collision.
Each data processing scenario (same name as in Config.DP) is the key of the matrix: its value is a dictionary, containing the mapping between the PD and the AlCa producers.
There is still the possibility to define the list of AlCa producers to be run in Express and Prompt Reconstruction.